Refactor: isolate postgres code
Isolate database specific code in a file to be able to create new files,
one for each database.

As a side effect, the Query struct, used by both main.rs and postgres.rs
was moved to config.rs.

While it's a bit overkill 6 source files in a small project, it simplify
envolving the code.